backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
openai==1.14.2
python-dotenv==1.0.1
openpyxl==3.1.2
